+++ src/main/java/com/mraof/minestuck/block/machine/TotemLatheBlock.java
@@ -6,11 +6,14 @@ import com.mraof.minestuck.blockentity.ItemStackBlockEntity;
 import com.mraof.minestuck.blockentity.machine.TotemLatheBlockEntity;
 import com.mraof.minestuck.util.CustomVoxelShape;
 import com.mraof.minestuck.util.MSRotationUtil;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.world.Containers;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
@@ -27,8 +30,11 @@ import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 
 import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.Map;
 
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
 public class TotemLatheBlock extends MultiMachineBlock
 {
 	protected final Map<Direction, VoxelShape> shape;
@@ -180,6 +186,16 @@ public class TotemLatheBlock extends MultiMachineBlock
 		{
 			super.createBlockStateDefinition(builder);
 			builder.add(COUNT);
+		}
+		
+		@Override
+		public void onRemove(BlockState state, Level level, BlockPos pos, BlockState newState, boolean isMoving)
+		{
+			if(!newState.is(state.getBlock()) &&
+					level.getBlockEntity(pos) instanceof TotemLatheBlockEntity blockEntity)
+				blockEntity.dropItems();
+			
+			super.onRemove(state, level, pos, newState, isMoving);
 		}
 	}
 }